requirements.txt
nest_asyncio
requests
boto3
azure-storage-file-datalake
azure-identity